+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Texture Replacer: Update the window's target material when a material is selected in the project view.
+- Show a message when replacement targets are not found.
 
 ### Fixed